// packages/diff-viewer/src/components/code-panel/components/hunk-list/types.ts
import { LoadMoreLinesHandler } from '../../../diff-viewer/types'
import { FileDiff } from '../../../../models/FileDiff'
import { LinePair } from './models/LinePair'

export type DiffLineType = 'add' | 'delete' | 'context' | 'hunk' | 'empty'
export type DisplayType = 'split' | 'unified'
export type Side = 'left' | 'right'
export type HunkDirection = 'up' | 'down' | 'out' | 'in' | 'in_up' | 'in_down'

export interface HunkListProps {
  /** A unique identifier for the file viewer. */
  id?: string
  /** The file diff object. */
  file: FileDiff
  /** Number of lines to request when user clicks "load more". Defaults to 5. */
  maxLinesToFetch?: number

  /** Called when user requests to load (expand) more lines around a hunk. */
  onLoadMoreLines?: LoadMoreLinesHandler
}

export interface HunkListHeaderProps {
  /* Whether the hunk list is collapsed */
  isCollapsed: boolean
  /* Whether the hunk list is viewed */
  isViewed: boolean
  /* The file diff */
  file: FileDiff
  /* The file path */
  filePath: string
  /* The display mode */
  mode: 'split' | 'unified'
  /* Whether the lines are wrapped */
  wrapLines: boolean

  /* Toggles the collapsed state of the hunk list */
  toggleCollapsed: (collapsed: boolean) => void
  /* Toggles the viewed state of the hunk list */
  toggleViewed: (viewed: boolean) => void
  /* Toggles the wrap lines state of the hunk list */
  onWrapLinesChange: (wrapLines: boolean) => void
}

export interface UnifiedViewerProps {
  /** Lines to display */
  lines: LinePair[]
  /** Whether to wrap long lines or enable horizontal scrolling (defaults to true). */
  wrapLines?: boolean
  /** Whether the viewer is currently visible (for performance optimization) */
  visible?: boolean
  /** Number of lines to request when user clicks "load more". Defaults to 5. */
  loadMoreLinesCount?: number

  /** Called when user requests to load (expand) more lines around a hunk. */
  onLoadMoreLines?: (line: LinePair, direction: HunkDirection) => void
}

export interface SplitViewerProps {
  /** Lines to display */
  lines: LinePair[]
  /** Number of lines to request when user clicks "load more". Defaults to 5. */
  loadMoreLinesCount?: number

  /** Called when user requests to load (expand) more lines around a hunk. */
  onLoadMoreLines?: (line: LinePair, direction: HunkDirection) => void
}